refactor: Line <-> LineSections 일급 객체 적용
