+++ src/main/java/nextstep/line/application/dto/line/LineDto.java
@@ -35,6 +35,6 @@ public class LineDto {
     }
 
     public static LineDto from(Line line) {
-        return new LineDto(line.getId(), line.getName(), line.getColor(), StationDto.from(line.getAllStations()));
+        return new LineDto(line.getId(), line.getName(), line.getColor(), StationDto.from(line.getStations()));
     }
 }
