+++ src/test/java/nextstep/line/unit/LineTest.java
@@ -238,7 +238,7 @@ class LineTest {
             // When
             신분당선.addSection(추가구간_엔티티(역삼역_엔티티, 선릉역_엔티티));
             // Then
-            List<Station> stations = 신분당선.getAllStations();
+            List<Station> stations = 신분당선.getStations();
             Station expectedLastStation = stations.get(stations.size() - 1);
             assertEquals(expectedLastStation.getName(), "선릉역");
         }
@@ -253,7 +253,7 @@ class LineTest {
             // When
             Line 신분당선 = 신분당선_엔티티(강남역_엔티티, 역삼역_엔티티);
             // Then
-            List<String> stationNames = 신분당선.getAllStations()
+            List<String> stationNames = 신분당선.getStations()
                     .stream().map(Station::getName)
                     .collect(Collectors.toList());
             assertEquals(stationNames, List.of("강남역", "역삼역"));
@@ -272,7 +272,7 @@ class LineTest {
             // When
             신분당선.addSection(추가구간_엔티티(역삼역_엔티티, 선릉역_엔티티));
             // Then
-            List<String> stationNames = 신분당선.getAllStations()
+            List<String> stationNames = 신분당선.getStations()
                     .stream().map(Station::getName)
                     .collect(Collectors.toList());
             assertEquals(stationNames, List.of("강남역", "역삼역", "선릉역"));
@@ -293,10 +293,10 @@ class LineTest {
             Line 신분당선 = 신분당선_엔티티(강남역_엔티티, 역삼역_엔티티);
             신분당선.addSection(추가구간_엔티티(역삼역_엔티티, 선릉역_엔티티));
             // When
-            신분당선.deleteStation(선릉역_엔티티.getId());
+            신분당선.deleteSection(선릉역_엔티티.getId());
             // Then
             assertEquals(신분당선.getSections().size(), 1);
-            assertFalse(신분당선.getAllStations().contains(선릉역_엔티티));
+            assertFalse(신분당선.getStations().contains(선릉역_엔티티));
         }
 
         /**
@@ -311,10 +311,10 @@ class LineTest {
             Line 신분당선 = 신분당선_엔티티(강남역_엔티티, 역삼역_엔티티);
             신분당선.addSection(추가구간_엔티티(역삼역_엔티티, 선릉역_엔티티));
             // When
-            신분당선.deleteStation(강남역_엔티티.getId());
+            신분당선.deleteSection(강남역_엔티티.getId());
             // Then
             assertEquals(신분당선.getSections().size(), 1);
-            assertFalse(신분당선.getAllStations().contains(강남역_엔티티));
+            assertFalse(신분당선.getStations().contains(강남역_엔티티));
         }
 
         /**
@@ -329,10 +329,10 @@ class LineTest {
             Line 신분당선 = 신분당선_엔티티(강남역_엔티티, 역삼역_엔티티);
             신분당선.addSection(추가구간_엔티티(역삼역_엔티티, 선릉역_엔티티));
             // When
-            신분당선.deleteStation(역삼역_엔티티.getId());
+            신분당선.deleteSection(역삼역_엔티티.getId());
             // Then
             assertEquals(신분당선.getSections().size(), 1);
-            assertFalse(신분당선.getAllStations().contains(역삼역_엔티티));
+            assertFalse(신분당선.getStations().contains(역삼역_엔티티));
         }
 
         /**
@@ -350,7 +350,7 @@ class LineTest {
             // Then
             assertThrows(
                     SectionException.class,
-                    () -> 신분당선.deleteStation(삼성역_엔티티.getId()),
+                    () -> 신분당선.deleteSection(삼성역_엔티티.getId()),
                     "should throw"
             );
         }
@@ -368,7 +368,7 @@ class LineTest {
             // Then
             assertThrows(
                     SectionException.class,
-                    () -> 신분당선.deleteStation(역삼역_엔티티.getId()),
+                    () -> 신분당선.deleteSection(역삼역_엔티티.getId()),
                     "should throw"
             );
